Moved config.ts
